backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
duckdb==0.9.2
networkx==3.2.1
dowhy==0.10.1
langchain==0.1.4
langgraph==0.0.21
groq==0.4.2
huggingface-hub==0.20.3
python-dotenv==1.0.1
pydantic==2.5.3
chromadb==0.4.22
pandas==2.1.4
sse-starlette==2.0.0
